Improve Authentication Service with sign-in/sign-out
The authentication service has been enhanced to support sign-in and sign-out features. Additional observables have been created to keep track of the logged in user's state. In the process, the method for adding users to the database was revised to ensure user data is stored according to their unique ID. Furthermore, the app component now executes login and logout functions.